+++ hydro_deploy/hydro_cli_examples/examples/panic_program/main.rs
@@ -1,0 +1,11 @@
+use std::io::Write;
+
+#[hydroflow::main]
+async fn main() {
+    let _ = hydroflow::util::cli::init::<()>().await;
+    println!("hello!");
+
+    std::io::stdout().flush().unwrap();
+
+    panic!();
+}
